requirements.txt
streamlit
python-dotenv

langchain
langchain-community
langchain-openai
langchain-chroma

chromadb
pypdf
tiktoken